Backend:mark assignments as checked & merge AI with submissions
