+++ server/src/routes/assigment.router.ts
@@ -1,12 +1,8 @@
 import { Router } from "express";
-import { getAssignmentsByRoomId } from "../controllers/assigments/assignment.controller";
-import { getStudentsSubmissionById } from "../controllers/assigments/getAssignmentwithAi.controller";
+import { getStudentsSubmissionById } from "../controllers/assigments/getStudentSubmissionByAssignId.controller";
 
 
 const router = Router();
-
-// GET /room/:roomId/assignments
-router.get("/room/:roomId/assignments", getAssignmentsByRoomId);
 
 router.get(
   "/submissions/:assignmentId/:studentId",
